Password Test cases, encryption, User guide and dev guide.

+++ src/main/java/seedu/address/storage/SecurityManager.java
@@ -5,40 +5,58 @@ import static java.util.Objects.requireNonNull;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.util.Base64;
 
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.events.ui.PasswordAcceptedEvent;
+import seedu.address.commons.exceptions.WrongPasswordException;
 import seedu.address.commons.util.FileUtil;
 
 //@@author syy94
+
 /**
  * Accesses and edits the password.
  */
 public class SecurityManager {
     private static String path = "data/pass";
 
-    public static void setPath(String path) {
+    public static void setPasswordStoragePath(String path) {
         SecurityManager.path = path;
     }
 
     /**
      * Saves password
      */
-    public static void savePass(String pass) throws IOException {
+    public static void savePass(String pass) throws IOException, NoSuchAlgorithmException {
         requireNonNull(pass);
         final File passFile = new File(path);
 
         final String trimmedPass = pass.trim();
+
         if (trimmedPass.length() != 0) {
+            final String encodedPass = encryptPass(trimmedPass);
             FileUtil.createIfMissing(passFile);
-            FileUtil.writeToFile(passFile, trimmedPass);
+            FileUtil.writeToFile(passFile, encodedPass);
         }
+    }
+
+    /**
+     * Simple Hashing of the password
+     */
+    private static String encryptPass(String pass) throws NoSuchAlgorithmException {
+        MessageDigest digest = MessageDigest.getInstance("SHA-256");
+        byte[] hash = digest.digest(pass.getBytes(StandardCharsets.UTF_8));
+
+        return Base64.getEncoder().encodeToString(hash);
     }
 
     /**
      * Checks if given password is correct and posts and event to enable the application features.
      */
-    public static boolean unlock(String pass) throws IOException {
+    public static boolean unlock(String pass) throws IOException, NoSuchAlgorithmException {
         boolean isUnlocked = checkPass(pass);
         if (isUnlocked) {
             EventsCenter.getInstance().post(new PasswordAcceptedEvent());
@@ -49,17 +67,22 @@ public class SecurityManager {
     /**
      * Checks if given password is correct.
      */
-    public static boolean checkPass(String pass) throws IOException {
-        return getPass().equals(pass);
+    public static boolean checkPass(String pass) throws IOException, NoSuchAlgorithmException {
+        return getPass().equals(encryptPass(pass));
     }
 
     /**
-     * removes the need for password on application start.
+     * Removes the need for password on application start after checking if the user knows the existing password.
      */
-    public static void removePass() {
-        final File passFile = new File(path);
-        if (FileUtil.isFileExists(passFile)) {
-            passFile.delete();
+    public static void removePass(String existingPass) throws IOException, NoSuchAlgorithmException,
+            WrongPasswordException {
+        if (checkPass(existingPass)) {
+            final File passFile = new File(path);
+            if (FileUtil.isFileExists(passFile)) {
+                passFile.delete();
+            }
+        } else {
+            throw new WrongPasswordException();
         }
     }
 
@@ -74,8 +97,6 @@ public class SecurityManager {
 
     /**
      * Retrieves the Password.
-     *
-     * @throws IOException
      */
     private static String getPass() throws IOException {
         final File file = new File(path);
